Moved createMqConfig into types and fixed integration test config
