+++ incoming-message-handler/src/index.ts
@@ -1,8 +1,7 @@
 import { SQSEvent } from "aws-lambda"
 import { isError } from "@handlers/common"
-import { createDynamoDbConfig } from "./types/DynamoDbConfig"
+import { createDynamoDbConfig, createMqConfig } from "./types"
 import MqGateway from "./gateways/MqGateway"
-import createMqConfig from "./utils/createMqConfig"
 import IncomingMessage from "./entities/IncomingMessage"
 import IncomingMessageDynamoGateway from "./gateways/IncomingMessageDynamoGateway"
 import SendRecordsUseCase from "./use-cases/SendRecordsUseCase"
