+++ incoming-message-handler/src/types/DynamoDbConfig.ts
