--- incoming-message-handler/src/types/DynamoDbConfig/createDynamoDbConfig.ts
@@ -1,8 +1,0 @@
-import DynamoDbConfig from "./DynamoDbConfig"
-
-export default function createDynamoDbConfig(): DynamoDbConfig {
-  return {
-    DYNAMO_URL: process.env.DYNAMO_URL,
-    DYNAMO_REGION: process.env.AWS_REGION
-  }
-}
