--- incoming-message-handler/src/types/DynamoDbConfig/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./DynamoDbConfig"
-export { default as createDynamoDbConfig } from "./createDynamoDbConfig"
