+++ incoming-message-handler/src/types/createDynamoDbConfig.ts
@@ -1,0 +1,8 @@
+import DynamoDbConfig from "./DynamoDbConfig"
+
+export default function createDynamoDbConfig(): DynamoDbConfig {
+  return {
+    DYNAMO_URL: process.env.DYNAMO_URL || `http://${process.env.LOCALSTACK_HOSTNAME}:4566`,
+    DYNAMO_REGION: process.env.AWS_REGION || "us-east-1"
+  }
+}
