+++ incoming-message-handler/src/types/index.ts
@@ -1,3 +1,5 @@
 export { default as DynamoDbConfig } from "./DynamoDbConfig"
+export { default as createDynamoDbConfig } from "./createDynamoDbConfig"
 export { default as MqConfig } from "./MqConfig"
+export { default as createMqConfig } from "./createMqConfig"
 export { default as MessageData } from "./MessageData"
